Improve participants form ui
