+++ src/features/participants/components/table/index.ts
@@ -1,0 +1,5 @@
+export { ActionCell } from "./action-cell";
+export { AddParticipantDialog } from "./add-participant-dialog";
+export { BulkDeleteButton } from "./bulk-delete-button";
+export { getColumns } from "./columns";
+export { ParticipantTableFilters } from "./filters";
